+++ CHANGELOG.md
@@ -7,10 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/atomist/cli/compare/0.1.0...HEAD)
 
-## [0.1.0](https://github.com/atomist/cli/tree/0.1.0) - 2018-07-02
+### Added
 
-Initial release
+-   Provide `--atomist-token` command-line option.
+
+### Changed
+
+-   Updated command-line options to latest terminology.
+
+### Removed
+
+-   SDM configuration helpers no longer necessary.
+
+## [0.1.0](https://github.com/atomist/cli/tree/0.1.0) - 2018-07-06
 
 ### Added
 
--   Atomist CLI, migrated from @atomist/automation-client
+-   Atomist CLI, migrated from @atomist/automation-client.